backend/requirements.txt
ag-ui-protocol==0.1.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
argcomplete==3.6.3
attrs==25.4.0
Authlib==1.6.6
beartype==0.22.9
boto3==1.42.30
botocore==1.42.30
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cohere==5.20.1
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.5.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fakeredis==2.33.0
fastapi==0.128.0
fastavro==1.12.1
fastmcp==2.14.3
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
genai-prices==0.0.51
google-auth==2.47.0
google-genai==1.59.0
googleapis-common-protos==1.72.0
griffe==1.15.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
logfire==4.19.0
logfire-api==4.19.0
lupa==2.6
markdown-it-py==4.0.0
mcp==1.25.0
mdurl==0.1.2
mistralai==1.9.11
more-itertools==10.8.0
multidict==6.7.0
nexus-rpc==1.2.0
openai==2.15.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
platformdirs==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-ai==1.44.0
pydantic-ai-slim==1.44.0
pydantic-evals==1.44.0
pydantic-graph==1.44.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.6
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.2.0
starlette==0.50.0
temporalio==1.20.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer==0.21.1
types-protobuf==6.32.1.20251210
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
